client: fix resource group controller cherry-pick
